Add files via upload
Pashupati is our way to lend maximum help to stray animals. For this we have designed a website which give a user interface for a person to either volunteer or donate to his will. "payment.html" is the payment gateway page we have made by ourselves which can be accessed through the design of the page of "Pashupati_project.html". We used HTML, CSS and Bootstrap to build this website. We hope our effort will make a difference and give stray animals a happy life.